Implement `Array2D` for non-owning 2D views of Host/Device memory (#14)
## Summary of Changes

This PR introduces a new `Array2D` class for managing 2D views into
contiguous arrays in host or device memory (powered by
[mdspan](https://en.cppreference.com/w/cpp/container/mdspan.html)). Note
that unlike the typical conainer types, `Array2D` does not own the
underlying buffer: it simply takes in a pointer and construct a 2D view
of it. As such, it does not deallocate the underlying buffer when the
object itself is out of scope. This is an intentional design because we
often need to pass the "views" of the buffer around within device code,
and we often need to keep many copies of the same view around within
device (in parallel), but we can only deallocate once on the host side.

Aside: while implementing this PR, I found CUDA std actually has an
undocumented
[submdspan](https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2023/p2630r4.html#design-of-submdspan)
type (at `cuda::std::submdspan`). This could potentially be useful in
creating multi-dimensional sub-views/non-contiguous views of the buffer
as we start to implement tiling.

## Test Plan

To run the unit tests included in the file:

```bash
pixi run ctest
```