Add Array2D bindings & support conversion to/from numpy arrays (#20)
(Part of MET-14)

## Summary of Changes

Introduce bindings for the `Array2D` class to facilitate data exchange
with numpy arrays. This includes methods for creating an Array2D from a
numpy array and converting it back, along with tests to validate
functionality and error handling.

Since `Array2D` are non-owning views of the underlying buffer, any
changes made on the new view will be visible to the original array.

## Test Plans

```bash
pixi run pytest
```