Implement Camera Intrinsics & compute ray direction in camera frame (#22)
(Part of MET-48)

## Summary of Changes

This PR complete the camera `Intrinsics` and defines two functions for
it:

- `get_ray_direction`: given a pixel coordinate `(px, py)`, returns the
direction of the ray in camera frame
- `get_ray_directions`: returns a 2D array of all ray directions in the
camera frame. If a range is specified, then this method only compute a
subset of the rays.

Depends on how we implement the kernel at the end, we may not need both
functions, but let's still have them here now just in case...

## Test Plans

On Python side, you can find a basic correctness testing in
`test_camera.py`, which checks the value of our
`Intrinsics::get_ray_direction` against the ground truth fucntion
refactored from `run_fmb.py`

Due to the complexity of binding `Array2D<Vec3D, device>` to Python, I'm
leaving the testing of the batched `get_ray_directions` to C++ side,
where `test_camera.cu` only performs a basic sanity check to ensure that
the adjacent rays are pointing toward different direction (and of
course, that the function runs without error).

As always, you can run the tests with

```bash
pixi run test
```