`Image` and `ImageView` classes (#26)
Closes MET-31.

## Summary of Changes

This PR adds `Image` and `ImageView` classes:

- `Image` is a host-side-only object that allocate & deallocate the
memory buffers in device/host automatically via RAII.
- `ImageView` is a non-owning view of `Image` which can be copy & pass
around on either device or host (or between them).

I've added the basic Python bindings for both classes on both device
types as well.

## Test Plan

The cpp test (`test_image.cu`) illustrates how the two classes are
expected to be used, checks that we can modify the underlying image
buffer through `ImageView`, and shows that we can copy `Image` from
device to host.

The python test (`test_image.py`) is pretty minimal at this point. It
only checks that the bindings are created successfully, and that we can
get the `confidence` and `depth` fields as 2D JAX arrays. This interface
can be used to plot & visualize the image data as we start putting
things together.

As always, you can run the included tests with

```bash
pixi run test
```