Refactor to use `Vec3D` instead of `Ray` (#29)
Part of MET-36

## Summary of Changes

Replace all instances of the `Ray` class with `Vec3D` (primarily in
intersector and getter) to streamline the code and remove unused
components.

This is part of my sequence of PRs for the forward kernels. The hope is
to keep each of them small so that even if we are not going to review it
now, it will be easier to catch up after the final deadline and
understand what changes were maded.


## Test Plan

```bash
pixi run test
```