Implement `get_pixel_coords` utility for forward kernel (#30)
Part of MET-36

## Summary of Changes

This PR implements the `get_pixel_coords` utility function, which will
be used in the forward kernel to compute the subset of pixel coordinates
that each thread is responsible for based on the thread & block IDs. As
part of this, I'm also adding a `int_ceil_div` utility (essentially
`int(ceil(a/b))`) to handle the potential edge cases where the number of
pixels isn't divisible by the number of threads.

## Test Plans

I've added some basic tests to check that `get_pixel_coords` guarantees
that each thread is covered by at least one thread (though it might be
possible for multiple threads to write to the same pixel when the job is
not divisible). To check them out, run

```bash
pixi run ctest
```